+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All changes to the package starting with v2.8.1 will be logged here.
 
 #### Enhancements
 * Implement logic to catch and present user-friendly error if a tenant is under maintenance
+* `britive.users.minimized_user_details` method to get a summarized set of user attributes given a list of user ids
 
 #### Bug Fixes
 * None